requirements.txt
fastmcp
beautifulsoup4
dotenv
